Allow usage in a wasm-web context
This replaces the use of parking_lot::Mutex with just plain
std::sync::Mutex due to
https://github.com/Amanieu/parking_lot/issues/269 and honestly the use
parking_lot was a bit of overkill here anyways.

This also introduces a `wasm-web` feature that can be enabled to enable
additional features in chrono and ring to allow tame-oauth to function
a wasm web (browser) context. We don't assume the user is in a web
context based on targetting wasm32-unknown-unknown, hence the feature.