Add support for EndUserCredentials.
Along the way to support #39, we need to support the gcloud
application_default_credentials.json / EndUserCredential format. The
flow is pretty similar to a ServiceAccount, but with a couple
differences:

 - You can't specify scopes dynamically (the scopes are assigned at
   authorization time, not access token / refresh time)

 - The endpoint isn't included (but oauth2.googleapis.com/token is the
   de facto endpoint)

 - After some period of time the refresh token is invalid, and you
   need to re-run `gcloud auth application-default login` (which sadly
   isn't check-able upfront, as the file doesn't store an
   `expires_in`) field.

I tested this out manually with a build of gcsfuser that does the full
Google default credentials flow.

Next up: I'll move that logic here into tame-oauth for everyone to use.