Add EndUserCredentials and implement get_default_google_credentials (#42)
* Add support for EndUserCredentials.

Along the way to support #39, we need to support the gcloud
application_default_credentials.json / EndUserCredential format. The
flow is pretty similar to a ServiceAccount, but with a couple
differences:

 - You can't specify scopes dynamically (the scopes are assigned at
   authorization time, not access token / refresh time)

 - The endpoint isn't included (but oauth2.googleapis.com/token is the
   de facto endpoint)

 - After some period of time the refresh token is invalid, and you
   need to re-run `gcloud auth application-default login` (which sadly
   isn't check-able upfront, as the file doesn't store an
   `expires_in`) field.

I tested this out manually with a build of gcsfuser that does the full
Google default credentials flow.

Next up: I'll move that logic here into tame-oauth for everyone to use.

* Implement get_default_google_credentials.

Now that we support all three flavors of OAuth credentials (Service
Account, gcloud ADC file, and Metadata server), we can do the full
"check this, then that, then that" flow that client libraries use (as
requested in #39).

With this change, getting the right `TokenProvider` is now just a call
to get_default_google_credentials(). Then the `get_token` and
`parse_token_response` work as expected.

This change does *not* yet handle caching. I think caching should move
out of the ServiceAccount handler and into its own generic "I can
cache anything" TokenProvider (CachedTokenProvider?).